requirements.txt
pandas==2.3.1
pyarrow==21.0.0
streamlit==1.48.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
dbt-postgres==1.9.0
